requirements.txt
asgiref==3.7.2
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
django-rest-passwordreset==1.3.0
django-mptt==0.15.0